+++ client/styles/components/component_kit/buttons.scss
@@ -1,105 +1,5 @@
 @import 'client/styles/facelift';
 
-.ExternalLinkElement {
-  display: flex;
-  max-width: fit-content;
-  font-weight: 500;
-  cursor: pointer;
-  .link-text {
-    color: $dark-purp;
-    margin-right: 15px;
-  }
-  &.button {
-    padding: 6px 24px 2px;
-    border: 1px solid $lite-purp;
-    border-radius: 10px;
-  }
-  &.inline {
-    padding: 3px 6px 0;
-    border-radius: 4px;
-  }
-  &:hover {
-    background-color: $lite-purp;
-  }
-}
-
-.RadioButton {
-  display: flex;
-  .radio-input {
-    display: flex;
-    margin-right: 15px;
-    input {
-      opacity: 0;
-      width: 0;
-      height: 0;
-    }
-  }
-  input {
-    display: none;
-    & + span {
-      display: inline-block;
-      position: relative;
-      padding-left: 20px;
-      cursor: pointer;
-      &:before {
-        content: '';
-        display: block;
-        position: absolute;
-        top: 0px;
-        left: 0px;
-        border-radius: 50%;
-        margin-right: 5px;
-        width: 16px;
-        height: 16px;
-        border: 1px solid $lite-gray;
-        background: $white;
-      }
-      &:after {
-        content: '';
-        display: block;
-        width: 10px;
-        height: 10px;
-        background: $purp-blue;
-        position: absolute;
-        border-radius: 50%;
-        top: 3px;
-        left: 3px;
-        opacity: 0;
-        transform: scale(0, 0);
-      }
-    }
-    &:hover + span:before {
-      border: 1px solid $purp-blue;
-      background: $purp-blue-lite;
-    }
-    &:hover + span:after {
-      opacity: 1;
-      transform: scale(1, 1);
-      background: $purp-blue;
-      top: 6px;
-      left: 6px;
-      width: 4px;
-      height: 4px;
-    }
-    &:checked + span:before {
-      background: $white;
-      border: 1px solid $purp-blue;
-    }
-    &:checked + span:after {
-      opacity: 1;
-      transform: scale(1, 1);
-      background: $purp-blue;
-      width: 10px;
-      height: 10px;
-      top: 3px;
-      left: 3px;
-    }
-  }
-  .radio-label {
-    @include body;
-  }
-}
-
 .RadioGroup {
   display: flex;
   flex-direction: column;
